[APM] Sort trace samples
