+++ x-pack/plugins/apm/server/routes/transactions/route.ts
@@ -16,13 +16,13 @@ import { setupRequest } from '../../lib/helpers/setup_request';
 import { getServiceTransactionGroups } from '../services/get_service_transaction_groups';
 import { getServiceTransactionGroupDetailedStatisticsPeriods } from '../services/get_service_transaction_group_detailed_statistics';
 import { getTransactionBreakdown } from './breakdown';
-import { getTransactionTraceSamples } from './trace_samples';
 import { getLatencyPeriods } from './get_latency_charts';
 import { getFailedTransactionRatePeriods } from './get_failed_transaction_rate_periods';
 import { getColdstartRatePeriods } from '../../lib/transaction_groups/get_coldstart_rate';
 import { createApmServerRoute } from '../apm_routes/create_apm_server_route';
 import { environmentRt, kueryRt, rangeRt } from '../default_api_types';
 import { offsetRt } from '../../../common/comparison_rt';
+import { getTraceSamples } from './trace_samples';
 
 const transactionGroupsMainStatisticsRoute = createApmServerRoute({
   endpoint:
@@ -314,7 +314,7 @@ const transactionTraceSamplesRoute = createApmServerRoute({
       end,
     } = params.query;
 
-    return getTransactionTraceSamples({
+    return getTraceSamples({
       environment,
       kuery,
       serviceName,
