+++ x-pack/plugins/apm/server/routes/transactions/trace_samples/index.ts
@@ -4,14 +4,27 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-
-import { Setup } from '../../../lib/helpers/setup_request';
-import { getTraceSamples } from './get_trace_samples';
+import { QueryDslQueryContainer } from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
+import { rangeQuery, kqlQuery } from '@kbn/observability-plugin/server';
+import { ProcessorEvent } from '@kbn/observability-plugin/common';
 import { withApmSpan } from '../../../utils/with_apm_span';
+import {
+  SERVICE_NAME,
+  TRACE_ID,
+  TRANSACTION_ID,
+  TRANSACTION_NAME,
+  TRANSACTION_SAMPLED,
+  TRANSACTION_TYPE,
+} from '../../../../common/elasticsearch_fieldnames';
+import { environmentQuery } from '../../../../common/utils/environment_query';
+import { Setup } from '../../../lib/helpers/setup_request';
+import { orderBy } from 'lodash';
 
-export async function getTransactionTraceSamples({
-  kuery,
+const TRACE_SAMPLES_SIZE = 500;
+
+export async function getTraceSamples({
   environment,
+  kuery,
   serviceName,
   transactionName,
   transactionType,
@@ -36,20 +49,62 @@ export async function getTransactionTraceSamples({
   start: number;
   end: number;
 }) {
-  return withApmSpan('get_transaction_trace_samples', async () => {
-    return await getTraceSamples({
-      environment,
-      kuery,
-      serviceName,
-      transactionName,
-      transactionType,
-      transactionId,
-      traceId,
-      sampleRangeFrom,
-      sampleRangeTo,
-      setup,
-      start,
-      end,
+  return withApmSpan('get_trace_samples', async () => {
+    const { apmEventClient } = setup;
+
+    const commonFilters = [
+      { term: { [SERVICE_NAME]: serviceName } },
+      { term: { [TRANSACTION_TYPE]: transactionType } },
+      { term: { [TRANSACTION_NAME]: transactionName } },
+      ...rangeQuery(start, end),
+      ...environmentQuery(environment),
+      ...kqlQuery(kuery),
+    ] as QueryDslQueryContainer[];
+
+    if (sampleRangeFrom !== undefined && sampleRangeTo !== undefined) {
+      commonFilters.push({
+        range: {
+          'transaction.duration.us': {
+            gte: sampleRangeFrom,
+            lte: sampleRangeTo,
+          },
+        },
+      });
+    }
+
+    const response = await apmEventClient.search('get_trace_samples_hits', {
+      apm: {
+        events: [ProcessorEvent.transaction],
+      },
+      _source: [TRANSACTION_ID, TRACE_ID, '@timestamp'],
+      body: {
+        query: {
+          bool: {
+            filter: [
+              ...commonFilters,
+              { term: { [TRANSACTION_SAMPLED]: true } },
+            ],
+            should: [
+              { term: { [TRACE_ID]: traceId } },
+              { term: { [TRANSACTION_ID]: transactionId } },
+            ] as QueryDslQueryContainer[],
+          },
+        },
+        size: TRACE_SAMPLES_SIZE,
+      },
     });
+
+    const traceSamples = orderBy(
+      response.hits.hits.map((hit) => ({
+        score: hit._score,
+        timestamp: hit._source['@timestamp'],
+        transactionId: hit._source.transaction.id,
+        traceId: hit._source.trace.id,
+      })),
+      ['score', 'timestamp'],
+      ['desc', 'desc']
+    );
+
+    return { traceSamples };
   });
 }
